Update tester agent
