+++ backend/app/agents/__init__.py
@@ -6,7 +6,7 @@ This module imports and exports all agent classes.
 from .base import Agent
 from .planner import PlannerAgent
 from .researcher import ResearchAgent
-from .code_generator import CodeGeneratorAgent
-from .test_executor import TestExecutionAgent
+from .developer import DeveloperAgent
+from .tester import TesterAgent
 
-__all__ = ['Agent', 'PlannerAgent', 'ResearchAgent', 'CodeGeneratorAgent', 'TestExecutionAgent']
+__all__ = ['Agent', 'PlannerAgent', 'ResearchAgent', 'DeveloperAgent', 'TesterAgent']
